chore: whisper request 패키지 변경
