+++ src/main/java/com/jolupbisang/demo/infrastructure/meeting/client/dto/request/ContextRequest.java
@@ -1,25 +1,20 @@
-package com.jolupbisang.demo.infrastructure.meeting.client.dto;
+package com.jolupbisang.demo.infrastructure.meeting.client.dto.request;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.Getter;
 import org.springframework.web.socket.TextMessage;
 import org.springframework.web.socket.BinaryMessage;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
-
 import java.io.IOException;
 
-@Getter
-public class ContextDoneRequest {
+public record ContextRequest(Dict dict) {
 
-    private final Dict dict;
-
-    private ContextDoneRequest(String groupId) {
-        this.dict = new Dict(WhisperRequestType.CONTEXT_DONE, groupId);
+    private ContextRequest(String groupId) {
+        this(new Dict(WhisperRequestType.CONTEXT, groupId));
     }
 
-    public static ContextDoneRequest of(String groupId) {
-        return new ContextDoneRequest(groupId);
+    public static ContextRequest of(String groupId) {
+        return new ContextRequest(groupId);
     }
 
     public TextMessage toTextMessage(ObjectMapper objectMapper) throws IOException {
@@ -38,14 +33,6 @@ public class ContextDoneRequest {
         return new BinaryMessage(buffer.array());
     }
 
-    @Getter
-    private static class Dict {
-        private final WhisperRequestType type;
-        private final String groupId;
-
-        public Dict(WhisperRequestType type, String groupId) {
-            this.type = type;
-            this.groupId = groupId;
-        }
+    private record Dict(WhisperRequestType type, String groupId) {
     }
 }
