+++ src/main/java/com/jolupbisang/demo/infrastructure/meeting/client/dto/request/MetaDataRequest.java
@@ -1,24 +1,18 @@
-package com.jolupbisang.demo.infrastructure.meeting.client.dto;
+package com.jolupbisang.demo.infrastructure.meeting.client.dto.request;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.Getter;
 import org.springframework.web.socket.TextMessage;
 import org.springframework.web.socket.BinaryMessage;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
-
 import java.io.IOException;
 import java.util.List;
 
-@Getter
-public class MetaDataRequest {
-
-    private final Dict dict;
-    private final MeetingMetaData metadata;
+public record MetaDataRequest(Dict dict, MeetingMetaData metadata) {
 
     private MetaDataRequest(String userId, String groupId, List<String> agenda, int numPeople, String meetingTopic) {
-        this.dict = new Dict(WhisperRequestType.METADATA, userId, groupId);
-        this.metadata = new MeetingMetaData(agenda, numPeople, meetingTopic);
+        this(new Dict(WhisperRequestType.METADATA, userId, groupId),
+             new MeetingMetaData(agenda, numPeople, meetingTopic));
     }
 
     public static MetaDataRequest of(String userId, String groupId, List<String> agenda, int numPeople, String meetingTopic) {
@@ -41,29 +35,9 @@ public class MetaDataRequest {
         return new BinaryMessage(buffer.array());
     }
 
-    @Getter
-    private static class Dict {
-        private final WhisperRequestType type;
-        private final String userId;
-        private final String groupId;
-
-        public Dict(WhisperRequestType type, String userId, String groupId) {
-            this.type = type;
-            this.userId = userId;
-            this.groupId = groupId;
-        }
+    private record Dict(WhisperRequestType type, String userId, String groupId) {
     }
 
-    @Getter
-    private static class MeetingMetaData {
-        private final int numPeople;
-        private final String meetingTopic;
-        private final List<String> agenda;
-
-        public MeetingMetaData(List<String> agenda, int numPeople, String meetingTopic) {
-            this.agenda = agenda;
-            this.numPeople = numPeople;
-            this.meetingTopic = meetingTopic;
-        }
+    private record MeetingMetaData(List<String> agenda, int numPeople, String meetingTopic) {
     }
 }
